fix(voice-recordings): stop pre-filling speaker ID for new dictionary entry
We have updated the voice recordings controller and view to remove the automatic pre-filling of the `speaker_id` when creating a new dictionary entry. Previously, the `speaker_id` was pre-populated with the ID of the last speaker from the recording's dictionary entries. However, this behavior has now been changed.

The `@recording.dictionary_entries.build` method in the voice_recordings_controller no longer receives the `speaker_id` as an argument. Instead, we now store the name of the last speaker in the `@last_speaker_name` variable, which is not used in the form construction but can be used elsewhere if needed.

In the corresponding view, the `speaker_id` text field within the form will not have a pre-filled value anymore, promoting manual entry or selection of the speaker's name. This adjustment ensures more accurate data entry and prevents the potential mistake of attributing a new entry to the previous speaker by default.

This change increases the flexibility of setting the speaker for a dictionary entry and avoids possible confusion or errors when speakers are different between entries.