feat(user-model): conditionally assign media_url to user hash
This commit introduces a conditional assignment of the `media_url` within the User model's data transformation logic. Previously, if a user had any recordings, the program would always assume there was a media attachment present and attempt to set the `media_url`. The update encapsulates this logic within an `if` statement, so now it only attempts to assign this URL if a sample is indeed present, thus preventing potential `nil` errors when accessing `media.url` on a nonexistent sample. This change enhances the robustness of the code when dealing with users who may not have associated recordings.